fix: 添加类型转换以解决 bindgen 跨平台类型不一致问题
bindgen 在不同平台上生成的 htmlParserOption 类型不一致：
- Linux/macOS/iOS: 生成 u32 类型的枚举值
- 但 htmlReadMemory 函数参数类型是 c_int (i32)

使用 htmlParserOption 类型别名进行位运算，在调用 C 函数时
转换为 i32，确保跨平台兼容性。

+++ xml-scraper/src/lib.rs
@@ -170,7 +170,7 @@ impl Document {
                 html.len() as i32,
                 ptr::null(),
                 c"UTF-8".as_ptr(),
-                raw_options,
+                raw_options as i32,
             );
 
             if doc_ptr.is_null() {
